requirements.txt
langchain==0.3.9
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langgraph==0.2.53
langgraph-checkpoint==2.0.7
langgraph-sdk==0.1.40
google-cloud-datastore==2.20.2 